Create transaction
