+++ components/Airdrop/StepJumioIframe/StepJumioIframe.tsx
@@ -1,6 +1,7 @@
 import clsx from 'clsx'
 import Loader from 'components/Loader'
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useRef } from 'react'
+import { useCreateKycFlow } from '../hooks/useCreateKycFlow'
 import { JumioFlowContainer } from '../JumioFlowContainer/JumioFlowContainer'
 import styles from './JumioIframe.module.css'
 
@@ -29,28 +30,6 @@ function useHandleJumioEvents(onFinish: () => void) {
   }, [onFinish])
 }
 
-function useJumioWebUrl(_userAddress: string) {
-  const [url, setUrl] = useState('')
-
-  useEffect(() => {
-    async function fetchJumioWebUrl() {
-      // const response = await fetch('/api/jumio/web')
-      // const data = await response.json()
-      const data = {
-        url: 'https://ironfish.web.amer-1.jumio.ai/web/v4/app?authorizationToken=eyJhbGciOiJIUzUxMiIsInppcCI6IkdaSVAifQ.H4sIAAAAAAAA_5XLOw4CMQwE0LukxtLaiT-ho6TlBoljnwAJJMTd2d0bMN2M3nxKvG_Pci0oqlJVsfNG5VKG-33tO0uvm9MAIqzQeDWYQQZ9KeOwtGQ--ImNJ0lWBs8W0IYY2PQAM6ctN1k1c8evjH-4PyJ33c8c_TxjTUHSgNCY0PpA6KgMywfNGDMlWvn-AEaqeFTgAAAA.TIeouQ2FOwhNvtW348y4prGWFL2ith9KoJqkKRTqxj2KJycdziXJxaCNEpSW1qvxNZA9RFDe0utH7zQn4GIshQ&locale=en',
-      }
-      setUrl(data.url)
-    }
-
-    fetchJumioWebUrl()
-  }, [])
-
-  return {
-    loading: url === '',
-    url,
-  }
-}
-
 type JumioIframeProps = {
   userAddress: string
   onFinish: () => void
@@ -62,7 +41,13 @@ export default function StepJumioIframe({
 }: JumioIframeProps) {
   useHandleJumioEvents(onFinish)
 
-  const { loading, url } = useJumioWebUrl(userAddress)
+  const createKycFlow = useCreateKycFlow(userAddress)
+
+  // eslint-disable-next-line
+  console.log({ createKycFlow })
+
+  const loading = true
+  const url = ''
 
   return (
     <JumioFlowContainer>
